+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/service/MultipleSubTasksService.java
@@ -1,6 +1,7 @@
 package de.codescape.jira.plugins.multiplesubtasks.service;
 
 import com.atlassian.jira.bc.issue.IssueService;
+import com.atlassian.jira.bc.user.search.AssigneeService;
 import com.atlassian.jira.config.PriorityManager;
 import com.atlassian.jira.config.SubTaskManager;
 import com.atlassian.jira.exception.CreateException;
@@ -12,7 +13,9 @@ import com.atlassian.jira.issue.issuetype.IssueType;
 import com.atlassian.jira.issue.priority.Priority;
 import com.atlassian.jira.project.Project;
 import com.atlassian.jira.security.JiraAuthenticationContext;
+import com.atlassian.jira.user.ApplicationUser;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
+import com.opensymphony.workflow.loader.ActionDescriptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -30,6 +33,7 @@ public class MultipleSubTasksService {
     private final IssueManager issueManager;
     private final SubTaskManager subTaskManager;
     private final PriorityManager priorityManager;
+    private final AssigneeService assigneeService;
     private final JiraAuthenticationContext jiraAuthenticationContext;
     private final SyntaxService syntaxService;
 
@@ -39,6 +43,7 @@ public class MultipleSubTasksService {
                                    @ComponentImport IssueManager issueManager,
                                    @ComponentImport SubTaskManager subTaskManager,
                                    @ComponentImport PriorityManager priorityManager,
+                                   @ComponentImport AssigneeService assigneeService,
                                    @ComponentImport JiraAuthenticationContext jiraAuthenticationContext,
                                    SyntaxService syntaxService) {
         this.issueService = issueService;
@@ -46,6 +51,7 @@ public class MultipleSubTasksService {
         this.issueManager = issueManager;
         this.subTaskManager = subTaskManager;
         this.priorityManager = priorityManager;
+        this.assigneeService = assigneeService;
         this.jiraAuthenticationContext = jiraAuthenticationContext;
         this.syntaxService = syntaxService;
     }
@@ -84,6 +90,7 @@ public class MultipleSubTasksService {
             // priority
             // try to find provided priority otherwise fall back to priority of parent issue
             if (subTaskRequest.getPriority() != null) {
+                // TODO a priority scheme can be configured per project
                 Priority priority = priorityManager.getPriorities().stream()
                     .filter(availablePriority -> availablePriority.getName().equals(subTaskRequest.getPriority()))
                     .findFirst()
@@ -103,6 +110,14 @@ public class MultipleSubTasksService {
             }
             if (newSubTask.getIssueType() == null) {
                 newSubTask.setIssueType(subTaskTypes.get(0));
+            }
+
+            // Assignee
+            // try to find provided assignee in the list of assignable users for current project and ignore users who are not assignable
+            if (subTaskRequest.getAssignee() != null) {
+                ApplicationUser assignee = assigneeService.findAssignableUsers(subTaskRequest.getAssignee(), projectObject).stream()
+                    .findFirst().orElse(null);
+                newSubTask.setAssignee(assignee);
             }
 
             // create and link the sub-task to the parent issue