+++ docs/changelog.md
@@ -47,6 +47,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: display input string and potential input errors after submit
 * feat: support explicit input for issue type of the created sub-tasks
 * chore: remove link to Travis CI
+* feat: support explicit input for assignee of the created sub-tasks
 
 [Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/1.0.0...HEAD
 [1.0.0]: https://github.com/codescape/jira-multiple-subtasks/tree/1.0.0